+++ docs/CHANGELOG.rst
@@ -11,6 +11,8 @@ Unreleased
 Fixed
 -----
 - Fix bug in ``get_or_run`` method
+- Fix a bug in ``QCTables`` where the removal of sample name suffixes
+  in MultiQC ``.txt`` files leads to incorrect parsing of MACS prepeak tables
 
 
 ===================